media: i2c: virt-cam: parse data lanes from dt
VCAM_LANES macro was defining the number of lanes which is not even
used while parsing that to DPHY parameters.

Instead of using macro for that it can be parsed from dt, where
that is already mentioned while connecting to the DPHY, also it
means driver compilation is not necessary.

Co-authored-by: shailparmar03 <shailparmar26@gmail.com>
Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>